section 6 webapi

+++ Vidly/Models/Movie.cs
@@ -29,6 +29,6 @@ namespace Vidly.Models
         [Required]
         [DefaultValue(0)]
         [Range(1,20, ErrorMessage ="The field Number in Stock must be between 1 and 20.")]
-        public byte Stock { get; set; }
+        public int Stock { get; set; }
     }
 }
